Split core.storage.models into domain modules
Convert the 2300-line models module into a package partitioned to match
core.storage.schemas. Public names are re-exported so existing imports
stay valid; schema (tables, columns, constraints, indexes) is unchanged.
