--- core/storage/models.py
@@ -1,2318 +1,0 @@
-"""
-SQLAlchemy Database Models for Vigil SOC
-
-Defines the database schema for cases, findings, and related entities.
-"""
-
-import uuid
-from datetime import datetime
-from typing import Any, List, Optional
-
-from sqlalchemy import (
-    ARRAY,
-    Boolean,
-    Column,
-    DateTime,
-    Float,
-    ForeignKey,
-    Index,
-    Integer,
-    Numeric,
-    String,
-    Table,
-    Text,
-    UniqueConstraint,
-    text,
-)
-from sqlalchemy.dialects.postgresql import JSONB
-from sqlalchemy.ext.mutable import MutableList
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
-
-from core.time import utcnow
-
-JSONBList = MutableList.as_mutable(JSONB)
-
-
-class Base(DeclarativeBase):
-    """Base class for all database models.
-
-    Overrides the declarative constructor for one reason: to refuse
-    ``metadata=``. SQLAlchemy accepts any kwarg for which ``hasattr(cls, key)``
-    holds, and ``metadata`` always holds — every declarative class inherits
-    ``Base.metadata``. The value lands on the instance, shadows the
-    ``MetaData``, reaches no column, and commits without error. Models that
-    need such a column rename it (``notification_metadata``,
-    ``decision_metadata``), so a bare ``metadata=`` is always a mistake, and
-    the only mistake here that nothing else can see. See #559.
-    """
-
-    def __init__(self, **kwargs: Any) -> None:
-        cls = type(self)
-        for key, value in kwargs.items():
-            if key == "metadata":
-                raise TypeError(
-                    f"{cls.__name__}(metadata=...) shadows the declarative "
-                    "MetaData and never reaches a column; pass the renamed "
-                    "column instead (e.g. notification_metadata)."
-                )
-            if not hasattr(cls, key):
-                raise TypeError(
-                    f"{key!r} is an invalid keyword argument for {cls.__name__}"
-                )
-            setattr(self, key, value)
-
-
-# Association table for case-finding many-to-many relationship
-case_findings = Table(
-    "case_findings",
-    Base.metadata,
-    Column(
-        "case_id",
-        String,
-        ForeignKey("cases.case_id", ondelete="CASCADE"),
-        primary_key=True,
-    ),
-    Column(
-        "finding_id",
-        String,
-        ForeignKey("findings.finding_id", ondelete="CASCADE"),
-        primary_key=True,
-    ),
-    Column("added_at", DateTime, default=utcnow, nullable=False),
-)
-
-
-class Finding(Base):
-    """Finding model - represents a security finding from DeepTempo LogLM."""
-
-    __tablename__ = "findings"
-
-    # Primary key
-    finding_id: Mapped[str] = mapped_column(String(50), primary_key=True)
-
-    anomaly_score: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-
-    # Human-readable description (populated from ingestion or synthesized from entity_context)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Entity context (optional fields)
-    entity_context: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Evidence links
-    evidence_links: Mapped[Optional[List[dict]]] = mapped_column(JSONB, nullable=True)
-
-    # Metadata
-    timestamp: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    data_source: Mapped[str] = mapped_column(String(50), nullable=False)
-    # Source-native ID. Combined with data_source it forms the dedup key
-    # for federated ingest (see uniq_findings_source_extid).
-    external_id: Mapped[Optional[str]] = mapped_column(String(255), nullable=True)
-    cluster_id: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
-    severity: Mapped[Optional[str]] = mapped_column(String(20), nullable=True)
-    status: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="new", server_default="new"
-    )
-
-    # AI-generated enrichment (cached analysis)
-    ai_enrichment: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Relationships
-    cases: Mapped[List["Case"]] = relationship(
-        "Case", secondary=case_findings, back_populates="findings", lazy="selectin"
-    )
-    mitre_prediction_rows: Mapped[List["FindingMitrePrediction"]] = relationship(
-        "FindingMitrePrediction",
-        back_populates="finding",
-        cascade="all, delete-orphan",
-        lazy="selectin",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_finding_timestamp", "timestamp"),
-        Index("idx_finding_severity", "severity"),
-        Index("idx_finding_status", "status"),
-        Index("idx_finding_data_source", "data_source"),
-        Index("idx_finding_cluster_id", "cluster_id"),
-        Index("idx_finding_anomaly_score", "anomaly_score"),
-        Index(
-            "idx_finding_description",
-            "description",
-            postgresql_ops={"description": "gin_trgm_ops"},
-            postgresql_using="gin",
-        ),
-        Index(
-            "uniq_findings_source_extid",
-            "data_source",
-            "external_id",
-            unique=True,
-            postgresql_where=text(
-                "data_source IS NOT NULL AND external_id IS NOT NULL"
-            ),
-        ),
-    )
-
-
-class FindingMitrePrediction(Base):
-    """One predicted ATT&CK technique (or ingest key) for a finding.
-
-    Keys are stored as text: parquet/CSV ingest uses tactic names, not only T-IDs.
-    """
-
-    __tablename__ = "finding_mitre_predictions"
-
-    finding_id: Mapped[str] = mapped_column(
-        String(50),
-        ForeignKey("findings.finding_id", ondelete="CASCADE"),
-        primary_key=True,
-    )
-    technique_id: Mapped[str] = mapped_column(Text, primary_key=True)
-    confidence: Mapped[float] = mapped_column(Float, nullable=False)
-
-    finding: Mapped["Finding"] = relationship(
-        "Finding", back_populates="mitre_prediction_rows"
-    )
-
-    __table_args__ = (
-        Index(
-            "idx_finding_mitre_predictions_technique_confidence",
-            "technique_id",
-            text("confidence DESC"),
-        ),
-    )
-
-
-class Case(Base):
-    """Case model - represents an investigation case grouping related findings."""
-
-    __tablename__ = "cases"
-
-    # Primary key
-    case_id: Mapped[str] = mapped_column(String(50), primary_key=True)
-
-    # Basic case information
-    title: Mapped[str] = mapped_column(String(200), nullable=False)
-    description: Mapped[str] = mapped_column(Text, nullable=True, default="")
-
-    # Status and priority
-    status: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="new", server_default="new"
-    )
-    priority: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="medium", server_default="medium"
-    )
-
-    # Assignment
-    assignee: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-
-    # Tags (array of strings)
-    tags: Mapped[List[str]] = mapped_column(ARRAY(String), nullable=True, default=list)
-
-    # Notes (JSONB array)
-    notes: Mapped[List[dict]] = mapped_column(JSONBList, nullable=True, default=list)
-
-    # Timeline events (JSONB array)
-    timeline: Mapped[List[dict]] = mapped_column(
-        JSONBList, nullable=False, default=list
-    )
-
-    # Activities (JSONB array)
-    activities: Mapped[Optional[List[dict]]] = mapped_column(
-        JSONBList, nullable=True, default=list
-    )
-
-    # Resolution steps (JSONB array)
-    resolution_steps: Mapped[Optional[List[dict]]] = mapped_column(
-        JSONBList, nullable=True, default=list
-    )
-
-    # MITRE ATT&CK techniques
-    mitre_techniques: Mapped[Optional[List[str]]] = mapped_column(
-        ARRAY(String), nullable=True
-    )
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Relationships
-    findings: Mapped[List[Finding]] = relationship(
-        "Finding", secondary=case_findings, back_populates="cases", lazy="selectin"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_status", "status"),
-        Index("idx_case_priority", "priority"),
-        Index("idx_case_assignee", "assignee"),
-        Index("idx_case_created_at", "created_at"),
-        Index("idx_case_updated_at", "updated_at"),
-    )
-
-
-class SketchMapping(Base):
-    """Timesketch mapping model - links cases/findings to Timesketch sketches."""
-
-    __tablename__ = "sketch_mappings"
-
-    # Primary key
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    # Mapping information
-    case_id: Mapped[Optional[str]] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=True
-    )
-    finding_id: Mapped[Optional[str]] = mapped_column(
-        String(50), ForeignKey("findings.finding_id", ondelete="CASCADE"), nullable=True
-    )
-
-    # Timesketch information
-    sketch_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    sketch_name: Mapped[str] = mapped_column(String(200), nullable=False)
-    sketch_url: Mapped[Optional[str]] = mapped_column(String(500), nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_sketch_case_id", "case_id"),
-        Index("idx_sketch_finding_id", "finding_id"),
-        Index("idx_sketch_id", "sketch_id"),
-    )
-
-
-class AttackLayer(Base):
-    """ATT&CK Navigator layer storage."""
-
-    __tablename__ = "attack_layers"
-
-    # Primary key
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    # Layer information
-    name: Mapped[str] = mapped_column(String(200), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    layer_data: Mapped[dict] = mapped_column(JSONB, nullable=False)
-
-    # Association with case (optional)
-    case_id: Mapped[Optional[str]] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="SET NULL"), nullable=True
-    )
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_attack_layer_case_id", "case_id"),
-        Index("idx_attack_layer_created_at", "created_at"),
-    )
-
-
-class AIDecisionLog(Base):
-    """
-    AI Decision Log - Tracks AI decisions for feedback and learning.
-
-    This model enables human oversight and continuous improvement of AI agents
-    by tracking all AI decisions, collecting human feedback, and measuring accuracy.
-    """
-
-    __tablename__ = "ai_decision_logs"
-
-    # Primary key
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    decision_id: Mapped[str] = mapped_column(String(50), unique=True, nullable=False)
-
-    # Decision context
-    agent_id: Mapped[str] = mapped_column(String(50), nullable=False, index=True)
-    workflow_id: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
-    finding_id: Mapped[Optional[str]] = mapped_column(
-        String(50), ForeignKey("findings.finding_id", ondelete="CASCADE"), nullable=True
-    )
-    case_id: Mapped[Optional[str]] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=True
-    )
-
-    # AI's decision
-    decision_type: Mapped[str] = mapped_column(String(50), nullable=False)
-    confidence_score: Mapped[float] = mapped_column(Float, nullable=False)
-    reasoning: Mapped[str] = mapped_column(Text, nullable=False)
-    recommended_action: Mapped[str] = mapped_column(Text, nullable=False)
-
-    # Additional decision metadata
-    decision_metadata: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Human feedback
-    human_reviewer: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    human_decision: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
-    feedback_comment: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Grading (0-1 scale)
-    accuracy_grade: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-    reasoning_grade: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-    action_appropriateness: Mapped[Optional[float]] = mapped_column(
-        Float, nullable=True
-    )
-
-    # Outcome tracking
-    actual_outcome: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
-    time_saved_minutes: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-
-    # Timestamps
-    timestamp: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    feedback_timestamp: Mapped[Optional[datetime]] = mapped_column(
-        DateTime, nullable=True
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_ai_decision_agent_id", "agent_id"),
-        Index("idx_ai_decision_finding_id", "finding_id"),
-        Index("idx_ai_decision_case_id", "case_id"),
-        Index("idx_ai_decision_timestamp", "timestamp"),
-        Index("idx_ai_decision_human_decision", "human_decision"),
-        Index("idx_ai_decision_actual_outcome", "actual_outcome"),
-    )
-
-
-class SystemConfig(Base):
-    """
-    System Configuration - Stores system-wide configuration settings.
-
-    This replaces file-based configs in ~/.vigil/ for better multi-user
-    support, ACID compliance, and audit trails.
-    """
-
-    __tablename__ = "system_config"
-
-    # Primary key
-    key: Mapped[str] = mapped_column(String(100), primary_key=True)
-
-    # Configuration value (flexible JSONB storage)
-    value: Mapped[dict] = mapped_column(JSONB, nullable=False)
-
-    # Metadata
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    config_type: Mapped[str] = mapped_column(
-        String(50), nullable=False, default="general", server_default="general"
-    )
-
-    # Audit fields
-    updated_by: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_system_config_type", "config_type"),
-        Index("idx_system_config_updated_at", "updated_at"),
-    )
-
-
-class UserPreference(Base):
-    """
-    User Preferences - Stores per-user preferences and settings.
-
-    Supports multi-user deployments with individual user settings.
-    """
-
-    __tablename__ = "user_preferences"
-
-    # Primary key
-    user_id: Mapped[str] = mapped_column(String(100), primary_key=True)
-
-    # Preferences as flexible JSONB
-    preferences: Mapped[dict] = mapped_column(JSONB, nullable=False, default={})
-
-    # User metadata
-    display_name: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    email: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Last login tracking
-    last_login: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-
-
-class IntegrationConfig(Base):
-    """
-    Integration Configuration - Stores non-sensitive integration settings.
-
-    Note: Secrets (API keys, passwords) remain in secrets_manager for security.
-    This stores connection details, preferences, and enabled/disabled state.
-    """
-
-    __tablename__ = "integration_configs"
-
-    # Primary key
-    integration_id: Mapped[str] = mapped_column(String(100), primary_key=True)
-
-    # Integration state
-    enabled: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-
-    # Configuration (non-sensitive only)
-    config: Mapped[dict] = mapped_column(JSONB, nullable=False, default={})
-
-    # Metadata
-    integration_name: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    integration_type: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Health status
-    last_test_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    last_test_success: Mapped[Optional[bool]] = mapped_column(Boolean, nullable=True)
-    last_error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Audit
-    updated_by: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_integration_enabled", "enabled"),
-        Index("idx_integration_type", "integration_type"),
-        Index("idx_integration_updated_at", "updated_at"),
-    )
-
-
-class FederationSource(Base):
-    """
-    Federation Source - Per-source state for the federated monitoring poller.
-
-    One row per data source the daemon pulls from on a configurable cadence.
-    Rows are auto-seeded on daemon boot from configured integrations (default
-    disabled). The global on/off lives in ``system_config`` under the key
-    ``federation.settings`` — a source only polls when both the global toggle
-    and its own ``enabled`` flag are true.
-    """
-
-    __tablename__ = "federation_sources"
-
-    # Primary key — matches integration ids (e.g. "splunk", "crowdstrike")
-    source_id: Mapped[str] = mapped_column(String(64), primary_key=True)
-
-    # Toggle + cadence
-    enabled: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=False, server_default="false"
-    )
-    interval_seconds: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=300, server_default="300"
-    )
-    max_items: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=100, server_default="100"
-    )
-
-    # Optional severity floor: only ingest findings >= this severity.
-    # Nullable means "no filter".
-    min_severity: Mapped[Optional[str]] = mapped_column(String(16), nullable=True)
-
-    # Adapter-defined cursor (e.g. {"earliest_time": "..."} for Splunk).
-    cursor: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default=text("'{}'::jsonb")
-    )
-
-    # Health
-    last_poll_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    last_success_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    last_error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    consecutive_errors: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0, server_default="0"
-    )
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default=text("now()")
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default=text("now()"),
-    )
-
-    __table_args__ = (Index("idx_federation_sources_enabled", "enabled"),)
-
-
-class ConfigAuditLog(Base):
-    """
-    Configuration Audit Log - Tracks all configuration changes for compliance.
-
-    Provides full audit trail of who changed what and when.
-    """
-
-    __tablename__ = "config_audit_log"
-
-    # Primary key
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    # What was changed
-    config_type: Mapped[str] = mapped_column(String(50), nullable=False)
-    config_key: Mapped[str] = mapped_column(String(200), nullable=False)
-
-    # Change details
-    action: Mapped[str] = mapped_column(
-        String(20), nullable=False
-    )  # create, update, delete
-    old_value: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-    new_value: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Who made the change
-    changed_by: Mapped[str] = mapped_column(String(100), nullable=False)
-    change_reason: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # When
-    timestamp: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_audit_config_type", "config_type"),
-        Index("idx_audit_config_key", "config_key"),
-        Index("idx_audit_changed_by", "changed_by"),
-        Index("idx_audit_timestamp", "timestamp"),
-    )
-
-
-# =============================================================================
-# Enhanced Case Management Models
-# =============================================================================
-
-
-class SLAPolicy(Base):
-    """
-    SLA Policy - Configurable service level agreement policies.
-
-    Defines response and resolution time requirements based on case priority.
-    """
-
-    __tablename__ = "sla_policies"
-
-    # Primary key
-    policy_id: Mapped[str] = mapped_column(String(50), primary_key=True)
-
-    # Policy details
-    name: Mapped[str] = mapped_column(String(100), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    priority_level: Mapped[str] = mapped_column(
-        String(20), nullable=False
-    )  # critical, high, medium, low
-
-    # Time requirements (in hours)
-    response_time_hours: Mapped[float] = mapped_column(Float, nullable=False)
-    resolution_time_hours: Mapped[float] = mapped_column(Float, nullable=False)
-
-    # Business hours settings
-    business_hours_only: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=True
-    )
-
-    # Escalation rules (JSONB)
-    escalation_rules: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Notification thresholds (e.g., [75, 90, 100] for 75%, 90%, 100% of time elapsed)
-    notification_thresholds: Mapped[Optional[List[int]]] = mapped_column(
-        ARRAY(Integer), nullable=True
-    )
-
-    # Status
-    is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    is_default: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_sla_policy_priority", "priority_level"),
-        Index("idx_sla_policy_active", "is_active"),
-        Index("idx_sla_policy_default", "is_default"),
-    )
-
-
-class CaseSLA(Base):
-    """
-    Case SLA - Tracks SLA compliance for individual cases.
-
-    Links cases to SLA policies and tracks deadlines, pauses, and breaches.
-    """
-
-    __tablename__ = "case_slas"
-
-    # Primary key
-    sla_id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-    sla_policy_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("sla_policies.policy_id"), nullable=False
-    )
-
-    # Deadlines
-    response_due: Mapped[datetime] = mapped_column(DateTime, nullable=False)
-    resolution_due: Mapped[datetime] = mapped_column(DateTime, nullable=False)
-
-    # Response tracking
-    response_completed_at: Mapped[Optional[datetime]] = mapped_column(
-        DateTime, nullable=True
-    )
-    response_sla_met: Mapped[Optional[bool]] = mapped_column(Boolean, nullable=True)
-
-    # Resolution tracking
-    resolution_completed_at: Mapped[Optional[datetime]] = mapped_column(
-        DateTime, nullable=True
-    )
-    resolution_sla_met: Mapped[Optional[bool]] = mapped_column(Boolean, nullable=True)
-
-    # Breach information
-    breached: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    breach_time: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    breach_reason: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Pause tracking
-    is_paused: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    paused_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    resumed_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    total_pause_duration: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0
-    )  # in seconds
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_sla_case_id", "case_id"),
-        Index("idx_case_sla_policy_id", "sla_policy_id"),
-        Index("idx_case_sla_response_due", "response_due"),
-        Index("idx_case_sla_resolution_due", "resolution_due"),
-        Index("idx_case_sla_breached", "breached"),
-    )
-
-
-class CaseComment(Base):
-    """
-    Case Comment - Discussion threads for cases.
-
-    Supports threaded conversations, @mentions, and rich text.
-    """
-
-    __tablename__ = "case_comments"
-
-    # Primary key
-    comment_id: Mapped[int] = mapped_column(
-        Integer, primary_key=True, autoincrement=True
-    )
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-    parent_comment_id: Mapped[Optional[int]] = mapped_column(
-        Integer,
-        ForeignKey("case_comments.comment_id", ondelete="CASCADE"),
-        nullable=True,
-    )
-
-    # Comment content
-    author: Mapped[str] = mapped_column(String(100), nullable=False)
-    content: Mapped[str] = mapped_column(Text, nullable=False)
-
-    # Mentions (user IDs)
-    mentions: Mapped[Optional[List[str]]] = mapped_column(ARRAY(String), nullable=True)
-
-    # Attachments (references to attachment IDs)
-    attachment_ids: Mapped[Optional[List[int]]] = mapped_column(
-        ARRAY(Integer), nullable=True
-    )
-
-    # Metadata
-    is_edited: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    is_deleted: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_comment_case_id", "case_id"),
-        Index("idx_case_comment_author", "author"),
-        Index("idx_case_comment_parent_id", "parent_comment_id"),
-        Index("idx_case_comment_created_at", "created_at"),
-    )
-
-
-class CaseWatcher(Base):
-    """
-    Case Watcher - Tracks users who are watching/following cases.
-
-    Enables notification subscriptions for case updates.
-    """
-
-    __tablename__ = "case_watchers"
-
-    # Composite primary key
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), primary_key=True
-    )
-    user_id: Mapped[str] = mapped_column(String(100), primary_key=True)
-
-    # Notification preferences (JSONB)
-    notification_preferences: Mapped[Optional[dict]] = mapped_column(
-        JSONB, nullable=True
-    )
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_watcher_case_id", "case_id"),
-        Index("idx_case_watcher_user_id", "user_id"),
-    )
-
-
-class CaseEvidence(Base):
-    """
-    Case Evidence - Tracks evidence and artifacts for cases.
-
-    Maintains chain of custody and evidence metadata.
-    """
-
-    __tablename__ = "case_evidence"
-
-    # Primary key
-    evidence_id: Mapped[int] = mapped_column(
-        Integer, primary_key=True, autoincrement=True
-    )
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-
-    # Evidence details
-    evidence_type: Mapped[str] = mapped_column(
-        String(50), nullable=False
-    )  # file, log, network_capture, memory_dump, etc.
-    name: Mapped[str] = mapped_column(String(200), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # File information
-    file_path: Mapped[Optional[str]] = mapped_column(String(500), nullable=True)
-    file_size: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-    file_hash_md5: Mapped[Optional[str]] = mapped_column(String(32), nullable=True)
-    file_hash_sha256: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
-
-    # Source information
-    source: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    collected_by: Mapped[str] = mapped_column(String(100), nullable=False)
-    collected_at: Mapped[datetime] = mapped_column(DateTime, nullable=False)
-
-    # Chain of custody (JSONB array of custody entries)
-    chain_of_custody: Mapped[List[dict]] = mapped_column(
-        JSONBList, nullable=False, default=list
-    )
-
-    # Analysis results
-    analysis_results: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Tags
-    tags: Mapped[Optional[List[str]]] = mapped_column(ARRAY(String), nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_evidence_case_id", "case_id"),
-        Index("idx_case_evidence_type", "evidence_type"),
-        Index("idx_case_evidence_collected_by", "collected_by"),
-        Index("idx_case_evidence_collected_at", "collected_at"),
-    )
-
-
-class CaseIOC(Base):
-    """
-    Case IOC - Indicators of Compromise associated with cases.
-
-    Tracks malicious indicators (IPs, domains, hashes, etc.).
-    """
-
-    __tablename__ = "case_iocs"
-
-    # Primary key
-    ioc_id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-
-    # IOC details
-    ioc_type: Mapped[str] = mapped_column(
-        String(50), nullable=False
-    )  # ip, domain, hash, url, email, file_name, etc.
-    value: Mapped[str] = mapped_column(String(500), nullable=False)
-
-    # Threat information
-    threat_level: Mapped[Optional[str]] = mapped_column(
-        String(20), nullable=True
-    )  # critical, high, medium, low
-    confidence: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-
-    # Source information
-    source: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    first_seen: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    last_seen: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-
-    # Enrichment data from threat intel sources
-    enrichment_data: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-    reputation_score: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-
-    # Tags and context
-    tags: Mapped[Optional[List[str]]] = mapped_column(ARRAY(String), nullable=True)
-    context: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Status
-    is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    is_false_positive: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=False
-    )
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_ioc_case_id", "case_id"),
-        Index("idx_case_ioc_type", "ioc_type"),
-        Index("idx_case_ioc_value", "value"),
-        Index("idx_case_ioc_threat_level", "threat_level"),
-        Index("idx_case_ioc_is_active", "is_active"),
-    )
-
-
-class CaseTask(Base):
-    """
-    Case Task - Tasks and sub-tasks for case investigations.
-
-    Supports hierarchical task structure and checklists.
-    """
-
-    __tablename__ = "case_tasks"
-
-    # Primary key
-    task_id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-    parent_task_id: Mapped[Optional[int]] = mapped_column(
-        Integer, ForeignKey("case_tasks.task_id", ondelete="CASCADE"), nullable=True
-    )
-
-    # Task details
-    title: Mapped[str] = mapped_column(String(200), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Assignment and status
-    assignee: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    status: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="pending"
-    )  # pending, in_progress, completed, cancelled
-    priority: Mapped[str] = mapped_column(String(20), nullable=False, default="medium")
-
-    # Time tracking
-    due_date: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    completed_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    estimated_hours: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-    actual_hours: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-
-    # Checklist items (JSONB array)
-    checklist_items: Mapped[Optional[List[dict]]] = mapped_column(JSONB, nullable=True)
-
-    # Metadata
-    task_order: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_task_case_id", "case_id"),
-        Index("idx_case_task_parent_id", "parent_task_id"),
-        Index("idx_case_task_assignee", "assignee"),
-        Index("idx_case_task_status", "status"),
-        Index("idx_case_task_due_date", "due_date"),
-    )
-
-
-class CaseTemplate(Base):
-    """
-    Case Template - Reusable templates for common investigation types.
-
-    Includes pre-defined tasks, playbooks, and default settings.
-    """
-
-    __tablename__ = "case_templates"
-
-    # Primary key
-    template_id: Mapped[str] = mapped_column(String(50), primary_key=True)
-
-    # Template details
-    name: Mapped[str] = mapped_column(String(200), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    template_type: Mapped[str] = mapped_column(
-        String(50), nullable=False
-    )  # malware, phishing, data_exfiltration, etc.
-
-    # Default case settings
-    default_priority: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="medium"
-    )
-    default_status: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="open"
-    )
-    default_sla_policy_id: Mapped[Optional[str]] = mapped_column(
-        String(50), nullable=True
-    )
-
-    # Task templates (JSONB array)
-    task_templates: Mapped[List[dict]] = mapped_column(
-        JSONB, nullable=False, default=list
-    )
-
-    # Playbook steps (JSONB array)
-    playbook_steps: Mapped[Optional[List[dict]]] = mapped_column(JSONB, nullable=True)
-
-    # MITRE ATT&CK techniques
-    applicable_mitre_techniques: Mapped[Optional[List[str]]] = mapped_column(
-        ARRAY(String), nullable=True
-    )
-
-    # Template metadata
-    tags: Mapped[Optional[List[str]]] = mapped_column(ARRAY(String), nullable=True)
-    is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    usage_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_template_type", "template_type"),
-        Index("idx_case_template_active", "is_active"),
-        Index("idx_case_template_usage_count", "usage_count"),
-    )
-
-
-class CaseRelationship(Base):
-    """
-    Case Relationship - Links related cases together.
-
-    Supports various relationship types (duplicate, related, parent-child, etc.).
-    """
-
-    __tablename__ = "case_relationships"
-
-    # Primary key
-    relationship_id: Mapped[int] = mapped_column(
-        Integer, primary_key=True, autoincrement=True
-    )
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-    related_case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-
-    # Relationship type
-    relationship_type: Mapped[str] = mapped_column(
-        String(50), nullable=False
-    )  # duplicate, related, parent, child, blocks, blocked_by
-
-    # Metadata
-    created_by: Mapped[str] = mapped_column(String(100), nullable=False)
-    notes: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_relationship_case_id", "case_id"),
-        Index("idx_case_relationship_related_case_id", "related_case_id"),
-        Index("idx_case_relationship_type", "relationship_type"),
-    )
-
-
-class CaseMetrics(Base):
-    """
-    Case Metrics - Performance and time tracking metrics for cases.
-
-    Tracks key metrics like MTTD, MTTR, MTTA, etc.
-    """
-
-    __tablename__ = "case_metrics"
-
-    # Primary key (one-to-one with cases)
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), primary_key=True
-    )
-
-    # Time metrics (in seconds)
-    time_to_detect: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-    time_to_respond: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-    time_to_contain: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-    time_to_resolve: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-
-    # Work tracking
-    total_work_hours: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-    analyst_handoffs_count: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0
-    )
-
-    # SLA tracking
-    sla_met: Mapped[Optional[bool]] = mapped_column(Boolean, nullable=True)
-    response_sla_met: Mapped[Optional[bool]] = mapped_column(Boolean, nullable=True)
-    resolution_sla_met: Mapped[Optional[bool]] = mapped_column(Boolean, nullable=True)
-
-    # Activity metrics
-    comment_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    evidence_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    ioc_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    task_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-
-class CaseAttachment(Base):
-    """
-    Case Attachment - File attachments for cases.
-
-    Stores metadata for files uploaded to cases.
-    """
-
-    __tablename__ = "case_attachments"
-
-    # Primary key
-    attachment_id: Mapped[int] = mapped_column(
-        Integer, primary_key=True, autoincrement=True
-    )
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-
-    # File details
-    filename: Mapped[str] = mapped_column(String(255), nullable=False)
-    file_path: Mapped[str] = mapped_column(String(500), nullable=False)
-    file_type: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    file_size: Mapped[int] = mapped_column(Integer, nullable=False)
-    mime_type: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-
-    # Metadata
-    uploaded_by: Mapped[str] = mapped_column(String(100), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    tags: Mapped[Optional[List[str]]] = mapped_column(ARRAY(String), nullable=True)
-
-    # Security scan results
-    virus_scan_result: Mapped[Optional[str]] = mapped_column(
-        String(50), nullable=True
-    )  # clean, infected, suspicious, not_scanned
-    scan_details: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_attachment_case_id", "case_id"),
-        Index("idx_case_attachment_uploaded_by", "uploaded_by"),
-        Index("idx_case_attachment_created_at", "created_at"),
-    )
-
-
-class CaseClosureInfo(Base):
-    """
-    Case Closure Info - Detailed closure metadata for closed cases.
-
-    Captures root cause, lessons learned, and post-incident information.
-    """
-
-    __tablename__ = "case_closure_info"
-
-    # Primary key (one-to-one with cases)
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), primary_key=True
-    )
-
-    # Closure details
-    closure_category: Mapped[str] = mapped_column(
-        String(50), nullable=False
-    )  # resolved, false_positive, duplicate, unable_to_resolve, etc.
-
-    # Root cause analysis
-    root_cause: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    contributing_factors: Mapped[Optional[List[str]]] = mapped_column(
-        ARRAY(String), nullable=True
-    )
-
-    # Post-incident review
-    lessons_learned: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    recommendations: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    recurrence_prevention: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # False positive details
-    false_positive_reason: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Summary
-    executive_summary: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Closure metadata
-    closed_by: Mapped[str] = mapped_column(String(100), nullable=False)
-    closure_notes: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Timestamps
-    closed_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-
-class CaseEscalation(Base):
-    """
-    Case Escalation - Tracks escalations for cases.
-
-    Records when and why cases are escalated to higher tiers or management.
-    """
-
-    __tablename__ = "case_escalations"
-
-    # Primary key
-    escalation_id: Mapped[int] = mapped_column(
-        Integer, primary_key=True, autoincrement=True
-    )
-
-    # References
-    case_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=False
-    )
-
-    # Escalation details
-    escalated_from: Mapped[str] = mapped_column(String(100), nullable=False)
-    escalated_to: Mapped[str] = mapped_column(String(100), nullable=False)
-    reason: Mapped[str] = mapped_column(Text, nullable=False)
-    urgency_level: Mapped[str] = mapped_column(
-        String(20), nullable=False
-    )  # low, medium, high, critical
-
-    # Status tracking
-    status: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="pending"
-    )  # pending, acknowledged, resolved
-
-    # Timestamps
-    escalated_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    acknowledged_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    resolved_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-
-    # Resolution
-    resolution_notes: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_escalation_case_id", "case_id"),
-        Index("idx_case_escalation_escalated_to", "escalated_to"),
-        Index("idx_case_escalation_status", "status"),
-        Index("idx_case_escalation_escalated_at", "escalated_at"),
-    )
-
-
-class CaseAuditLog(Base):
-    """
-    Case Audit Log - Field-level audit trail for case changes.
-
-    Tracks all modifications to cases and related entities for compliance.
-    """
-
-    __tablename__ = "case_audit_logs"
-
-    # Primary key
-    audit_id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    # What was changed
-    entity_type: Mapped[str] = mapped_column(
-        String(50), nullable=False
-    )  # case, comment, evidence, ioc, etc.
-    entity_id: Mapped[str] = mapped_column(String(100), nullable=False)
-
-    # Change details
-    action: Mapped[str] = mapped_column(
-        String(20), nullable=False
-    )  # create, update, delete
-    field_name: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    old_value: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    new_value: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Additional context
-    change_summary: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    # Who made the change
-    changed_by: Mapped[str] = mapped_column(String(100), nullable=False)
-    user_agent: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    ip_address: Mapped[Optional[str]] = mapped_column(String(45), nullable=True)
-
-    # Timestamp
-    timestamp: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_audit_entity_type", "entity_type"),
-        Index("idx_case_audit_entity_id", "entity_id"),
-        Index("idx_case_audit_changed_by", "changed_by"),
-        Index("idx_case_audit_timestamp", "timestamp"),
-        Index("idx_case_audit_action", "action"),
-    )
-
-
-class User(Base):
-    """
-    User Model - System users with authentication and authorization.
-
-    Stores user credentials, profile information, and role assignments.
-    """
-
-    __tablename__ = "users"
-
-    # Primary key
-    user_id: Mapped[str] = mapped_column(String(50), primary_key=True)
-
-    # Authentication
-    username: Mapped[str] = mapped_column(String(100), unique=True, nullable=False)
-    email: Mapped[str] = mapped_column(String(200), unique=True, nullable=False)
-    password_hash: Mapped[str] = mapped_column(String(255), nullable=False)
-
-    # Profile
-    full_name: Mapped[str] = mapped_column(String(200), nullable=False)
-
-    # Role and permissions
-    role_id: Mapped[str] = mapped_column(
-        String(50), ForeignKey("roles.role_id"), nullable=False
-    )
-
-    # Status
-    is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    is_verified: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-
-    # MFA
-    mfa_enabled: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    mfa_secret: Mapped[Optional[str]] = mapped_column(String(255), nullable=True)
-    mfa_recovery_codes: Mapped[list] = mapped_column(
-        JSONB, nullable=False, default=list
-    )
-
-    # Session tracking
-    last_login: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    login_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
-    # Failed-login tracking and account lockout
-    failed_login_count: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0, server_default="0"
-    )
-    locked_until: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-
-    # Password history — list of prior bcrypt hashes, newest first. Used
-    # to reject reuse of the last N passwords. Capped in application code.
-    password_history: Mapped[List[str]] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-    password_changed_at: Mapped[Optional[datetime]] = mapped_column(
-        DateTime, nullable=True
-    )
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_user_username", "username"),
-        Index("idx_user_email", "email"),
-        Index("idx_user_role_id", "role_id"),
-        Index("idx_user_is_active", "is_active"),
-    )
-
-
-class Role(Base):
-    """
-    Role Model - Defines user roles and their permissions.
-
-    RBAC (Role-Based Access Control) system for authorization.
-    """
-
-    __tablename__ = "roles"
-
-    # Primary key
-    role_id: Mapped[str] = mapped_column(String(50), primary_key=True)
-
-    # Role details
-    name: Mapped[str] = mapped_column(String(100), unique=True, nullable=False)
-    description: Mapped[str] = mapped_column(Text, nullable=False)
-
-    # Permissions (JSONB for flexibility)
-    permissions: Mapped[dict] = mapped_column(JSONB, nullable=False, default={})
-
-    # System role flag (cannot be deleted/modified)
-    is_system_role: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    # Indexes
-    __table_args__ = (Index("idx_role_name", "name"),)
-
-
-# =============================================================================
-# Autonomous Orchestrator Models
-# =============================================================================
-
-
-class Investigation(Base):
-    """Tracks an autonomous investigation assignment managed by the orchestrator."""
-
-    __tablename__ = "investigations"
-
-    investigation_id: Mapped[str] = mapped_column(String(60), primary_key=True)
-    case_id: Mapped[Optional[str]] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="SET NULL"), nullable=True
-    )
-    workflow_id: Mapped[str] = mapped_column(String(50), nullable=False)
-
-    trigger_type: Mapped[str] = mapped_column(String(30), nullable=False)
-    # Finding ids, not objects (#554). Both writers build a list of
-    # ``findings.finding_id`` strings -- services/daemon/orchestrator.py:407
-    # (``[f.get("finding_id") for f in findings ...]``) and :942
-    # (``finding_ids[:10]``) -- and every reader treats the elements as those
-    # strings: services/api/routers/orchestrator.py:484 tests them for set
-    # membership against ``Finding.finding_id``, and orchestrator.py:1098 writes
-    # ``trigger_ids[0]`` straight into ``AIDecisionLog.finding_id``, a
-    # String(50) FK. The old ``List[dict]`` annotation matched no writer or
-    # reader. Storage stays JSONB; promoting the column is deferred to #468.
-    trigger_ids: Mapped[List[str]] = mapped_column(JSONB, nullable=False, default=list)
-
-    status: Mapped[str] = mapped_column(String(20), nullable=False, default="queued")
-
-    workdir: Mapped[str] = mapped_column(String(255), nullable=False)
-
-    current_step: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    total_steps: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
-    iteration_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    max_iterations: Mapped[int] = mapped_column(Integer, nullable=False, default=50)
-
-    input_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    output_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    cost_usd: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
-    max_cost_usd: Mapped[float] = mapped_column(Float, nullable=False, default=5.0)
-
-    priority: Mapped[str] = mapped_column(String(20), nullable=False, default="medium")
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    started_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    completed_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    last_activity_at: Mapped[Optional[datetime]] = mapped_column(
-        DateTime, nullable=True
-    )
-    max_runtime_seconds: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=3600
-    )
-
-    summary: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    proposed_actions: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-    master_review_notes: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    error_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    last_error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    current_activity: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    __table_args__ = (
-        Index("idx_investigation_status", "status"),
-        Index("idx_investigation_case_id", "case_id"),
-        Index("idx_investigation_priority", "priority"),
-        Index("idx_investigation_created_at", "created_at"),
-        Index("idx_investigation_workflow_id", "workflow_id"),
-    )
-
-
-class InvestigationLog(Base):
-    """Append-only audit log for investigation agent actions."""
-
-    __tablename__ = "investigation_logs"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    investigation_id: Mapped[str] = mapped_column(
-        String(60),
-        ForeignKey("investigations.investigation_id", ondelete="CASCADE"),
-        nullable=False,
-    )
-    timestamp: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    event_type: Mapped[str] = mapped_column(String(30), nullable=False)
-    details: Mapped[dict] = mapped_column(JSONB, nullable=False, default={})
-    tokens_used: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
-    __table_args__ = (
-        Index("idx_inv_log_investigation_id", "investigation_id"),
-        Index("idx_inv_log_timestamp", "timestamp"),
-        Index("idx_inv_log_event_type", "event_type"),
-    )
-
-
-class LLMInteractionLog(Base):
-    """Durable audit log of Claude API interactions for chain-of-thought visibility.
-
-    One row per Anthropic API response (per tool-use iteration). Captures
-    full untruncated thinking blocks, tool call chains, token usage, and
-    cost so analysts can audit AI decision-making post-hoc.
-    """
-
-    __tablename__ = "llm_interaction_logs"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    interaction_id: Mapped[str] = mapped_column(String(60), nullable=False, unique=True)
-    session_id: Mapped[Optional[str]] = mapped_column(String(120), nullable=True)
-    agent_id: Mapped[Optional[str]] = mapped_column(String(60), nullable=True)
-    investigation_id: Mapped[Optional[str]] = mapped_column(String(60), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    model: Mapped[str] = mapped_column(String(80), nullable=False)
-    request_messages: Mapped[list] = mapped_column(JSONB, nullable=False, default=list)
-    system_prompt: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    thinking_enabled: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=False
-    )
-    thinking_budget: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-    thinking_content: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    response_content: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    tool_calls: Mapped[list] = mapped_column(JSONB, nullable=False, default=list)
-    tool_results: Mapped[list] = mapped_column(JSONB, nullable=False, default=list)
-    stop_reason: Mapped[Optional[str]] = mapped_column(String(40), nullable=True)
-    input_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    output_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    cache_read_tokens: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    cache_creation_tokens: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0
-    )
-    cost_usd: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False, default=0)
-    duration_ms: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    # Bifrost virtual-key attribution (#186). Stores the VK the call was
-    # made under so we can group spend per-tenant once Vigil grows a
-    # tenant model. Empty / NULL for calls made before the budget feature
-    # was enabled or while running in DEV_MODE / LLM_BUDGET_UNLIMITED.
-    virtual_key_id: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
-
-    __table_args__ = (
-        Index("idx_llm_interaction_session", "session_id"),
-        Index("idx_llm_interaction_agent", "agent_id"),
-        Index("idx_llm_interaction_investigation", "investigation_id"),
-        Index("idx_llm_interaction_created", "created_at"),
-        Index("idx_llm_interaction_vk", "virtual_key_id", "created_at"),
-    )
-
-
-class SharedIOC(Base):
-    """Cross-investigation IOC index for deduplication and correlation."""
-
-    __tablename__ = "shared_iocs"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    investigation_id: Mapped[str] = mapped_column(
-        String(60),
-        ForeignKey("investigations.investigation_id", ondelete="CASCADE"),
-        nullable=False,
-    )
-    ioc_type: Mapped[str] = mapped_column(String(30), nullable=False)
-    value: Mapped[str] = mapped_column(String(500), nullable=False)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    __table_args__ = (
-        Index("idx_shared_ioc_value", "value"),
-        Index("idx_shared_ioc_type", "ioc_type"),
-        Index("idx_shared_ioc_investigation", "investigation_id"),
-    )
-
-
-class CaseNotification(Base):
-    """
-    Case Notification - Notification queue for case-related events.
-
-    Tracks notifications to be delivered to users about case updates.
-    """
-
-    __tablename__ = "case_notifications"
-
-    # Primary key
-    notification_id: Mapped[int] = mapped_column(
-        Integer, primary_key=True, autoincrement=True
-    )
-
-    # References
-    case_id: Mapped[Optional[str]] = mapped_column(
-        String(50), ForeignKey("cases.case_id", ondelete="CASCADE"), nullable=True
-    )
-    user_id: Mapped[str] = mapped_column(String(100), nullable=False)
-
-    # Notification details
-    notification_type: Mapped[str] = mapped_column(
-        String(50), nullable=False
-    )  # case_assigned, comment_mention, sla_warning, escalation, etc.
-    title: Mapped[str] = mapped_column(String(200), nullable=False)
-    message: Mapped[str] = mapped_column(Text, nullable=False)
-
-    # Delivery settings
-    delivery_channel: Mapped[str] = mapped_column(
-        String(20), nullable=False, default="ui"
-    )  # ui, email, slack, teams, pagerduty
-    priority: Mapped[str] = mapped_column(String(20), nullable=False, default="normal")
-
-    # Status
-    is_read: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    is_sent: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    sent_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    read_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-
-    # Metadata (renamed from 'metadata' to avoid SQLAlchemy reserved word conflict)
-    notification_metadata: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    # Indexes
-    __table_args__ = (
-        Index("idx_case_notification_case_id", "case_id"),
-        Index("idx_case_notification_user_id", "user_id"),
-        Index("idx_case_notification_type", "notification_type"),
-        Index("idx_case_notification_is_read", "is_read"),
-        Index("idx_case_notification_is_sent", "is_sent"),
-        Index("idx_case_notification_created_at", "created_at"),
-    )
-
-
-class CustomWorkflow(Base):
-    """
-    Custom Workflow Model - User-created multi-agent workflow definitions.
-
-    File-based WORKFLOW.md definitions remain supported separately. This table
-    holds workflows created/edited via the Workflow Builder UI.
-    """
-
-    __tablename__ = "custom_workflows"
-
-    workflow_id: Mapped[str] = mapped_column(String(100), primary_key=True)
-
-    name: Mapped[str] = mapped_column(String(200), nullable=False)
-    description: Mapped[str] = mapped_column(Text, nullable=False)
-    use_case: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    trigger_examples: Mapped[list] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-    phases: Mapped[list] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-    graph_layout: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default="{}"
-    )
-
-    is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    created_by: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    version: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        server_default="now()",
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    __table_args__ = (
-        Index("idx_custom_workflows_active", "is_active"),
-        Index("idx_custom_workflows_created_by", "created_by"),
-        Index("idx_custom_workflows_name", "name"),
-    )
-
-
-class WorkflowRun(Base):
-    """Per-invocation record of ``execute_workflow`` (#127).
-
-    One row per `/api/workflows/{id}/execute` call, used for history +
-    audit. The parent row always exists; `workflow_run_phases` rows are
-    reserved for phase-by-phase execution (#128) and may be absent.
-    """
-
-    __tablename__ = "workflow_runs"
-
-    run_id: Mapped[str] = mapped_column(String(80), primary_key=True)
-    workflow_id: Mapped[str] = mapped_column(Text, nullable=False)
-    workflow_version: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-    workflow_source: Mapped[str] = mapped_column(
-        String(16), nullable=False, default="file", server_default="file"
-    )
-    workflow_name: Mapped[str] = mapped_column(Text, nullable=False)
-
-    status: Mapped[str] = mapped_column(String(16), nullable=False)
-    triggered_by: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    trigger_context: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default="{}"
-    )
-
-    started_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    finished_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    duration_ms: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-
-    total_cost_usd: Mapped[float] = mapped_column(
-        Numeric(10, 4), nullable=False, default=0, server_default="0"
-    )
-    result_summary: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    skill_tools_available: Mapped[list] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-    error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    # Set when an operator removes the run from History. The row and its ledger stay.
-    deleted_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-
-    __table_args__ = (
-        Index("idx_workflow_runs_workflow_id", "workflow_id", "started_at"),
-        Index("idx_workflow_runs_started_at", "started_at"),
-    )
-
-
-class WorkflowRunPhase(Base):
-    """Per-phase record within a workflow run.
-
-    Reserved for phase-by-phase execution (#128). The table ships with
-    the schema so the audit story is complete, but no rows are written
-    until phase-level execution lands.
-    """
-
-    __tablename__ = "workflow_run_phases"
-
-    run_id: Mapped[str] = mapped_column(
-        String(80),
-        ForeignKey("workflow_runs.run_id", ondelete="CASCADE"),
-        primary_key=True,
-    )
-    phase_id: Mapped[str] = mapped_column(Text, primary_key=True)
-    phase_order: Mapped[int] = mapped_column(Integer, nullable=False)
-    agent_id: Mapped[str] = mapped_column(Text, nullable=False)
-    status: Mapped[str] = mapped_column(String(16), nullable=False)
-
-    started_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    finished_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    duration_ms: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
-
-    input_context: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default="{}"
-    )
-    output: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default="{}"
-    )
-    approval_state: Mapped[Optional[str]] = mapped_column(String(16), nullable=True)
-    cost_usd: Mapped[float] = mapped_column(
-        Numeric(10, 4), nullable=False, default=0, server_default="0"
-    )
-    error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    __table_args__ = (Index("idx_workflow_run_phases_run_id", "run_id", "phase_order"),)
-
-
-class ApprovalAction(Base):
-    """Pending human-in-the-loop approval (#128).
-
-    Supersedes the JSON-file persistence that ApprovalService used to
-    do. Workflow phase-level approvals link back to the paused run via
-    ``workflow_run_id`` + ``workflow_phase_id``. Non-workflow approvals
-    (daemon containment actions, etc.) leave those columns null.
-    """
-
-    __tablename__ = "approval_actions"
-
-    action_id: Mapped[str] = mapped_column(String(80), primary_key=True)
-    action_type: Mapped[str] = mapped_column(String(40), nullable=False)
-    title: Mapped[str] = mapped_column(Text, nullable=False)
-    description: Mapped[str] = mapped_column(Text, nullable=False)
-    target: Mapped[str] = mapped_column(Text, nullable=False)
-    confidence: Mapped[float] = mapped_column(
-        Numeric(4, 3), nullable=False, default=0, server_default="0"
-    )
-    reason: Mapped[str] = mapped_column(Text, nullable=False)
-    evidence: Mapped[list] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    created_by: Mapped[str] = mapped_column(String(100), nullable=False)
-    requires_approval: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=True, server_default="true"
-    )
-    status: Mapped[str] = mapped_column(String(16), nullable=False)
-    approved_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    approved_by: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    executed_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    execution_result: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-    rejection_reason: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    parameters: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default="{}"
-    )
-    workflow_run_id: Mapped[Optional[str]] = mapped_column(
-        String(80),
-        ForeignKey("workflow_runs.run_id", ondelete="SET NULL"),
-        nullable=True,
-    )
-    workflow_phase_id: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    __table_args__ = (
-        Index("idx_approval_actions_status_created", "status", "created_at"),
-        Index("idx_approval_actions_workflow_run", "workflow_run_id"),
-    )
-
-
-class Skill(Base):
-    """Skill model - reusable, parameterized SOC capability (detection,
-    enrichment, response, reporting) that agents and workflows can invoke."""
-
-    __tablename__ = "skills"
-
-    skill_id: Mapped[str] = mapped_column(String(32), primary_key=True)
-
-    name: Mapped[str] = mapped_column(String(255), nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    category: Mapped[str] = mapped_column(String(32), nullable=False)
-
-    # JSON Schema for skill parameters (the inputs the skill accepts).
-    input_schema: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default="{}"
-    )
-    # JSON Schema for skill output.
-    output_schema: Mapped[dict] = mapped_column(
-        JSONB, nullable=False, default=dict, server_default="{}"
-    )
-    # MCP tool names required by this skill
-    # (e.g. ["splunk.search", "virustotal.hash_lookup"]).
-    required_tools: Mapped[List[str]] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-    # LLM instructions; may contain {{param}} placeholders.
-    prompt_template: Mapped[str] = mapped_column(Text, nullable=False)
-    # Ordered execution steps (tool calls / prompts / transforms) — interpreted
-    # by the future skill-execution worker.
-    execution_steps: Mapped[List[dict]] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-
-    is_active: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=True, server_default="true"
-    )
-    created_by: Mapped[Optional[str]] = mapped_column(String(255), nullable=True)
-    version: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=1, server_default="1"
-    )
-
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        server_default="now()",
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    __table_args__ = (
-        Index("idx_skill_category", "category"),
-        Index("idx_skill_is_active", "is_active"),
-        Index(
-            "idx_skill_name_trgm",
-            "name",
-            postgresql_ops={"name": "gin_trgm_ops"},
-            postgresql_using="gin",
-        ),
-    )
-
-    @staticmethod
-    def generate_skill_id() -> str:
-        """Generate a new skill_id in the form s-YYYYMMDD-XXXXXXXX."""
-        ts = utcnow().strftime("%Y%m%d")
-        return f"s-{ts}-{uuid.uuid4().hex[:8].upper()}"
-
-
-class CustomAgent(Base):
-    """User-defined SOC agent created via the Agent Builder UI."""
-
-    __tablename__ = "custom_agents"
-
-    id: Mapped[str] = mapped_column(String, primary_key=True)
-    name: Mapped[str] = mapped_column(Text, nullable=False)
-    description: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    icon: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    color: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    specialization: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    role: Mapped[str] = mapped_column(Text, nullable=False)
-    extra_principles: Mapped[str] = mapped_column(
-        Text, nullable=False, default="", server_default=""
-    )
-    methodology: Mapped[str] = mapped_column(
-        Text, nullable=False, default="", server_default=""
-    )
-    system_prompt_override: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    recommended_tools: Mapped[list] = mapped_column(
-        JSONB, nullable=False, default=list, server_default="[]"
-    )
-    max_tokens: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=4096, server_default="4096"
-    )
-    enable_thinking: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=False, server_default="false"
-    )
-    model: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    component_category: Mapped[str] = mapped_column(
-        String(32),
-        nullable=False,
-        default="investigation",
-        server_default="investigation",
-    )
-    # Origin agent ID this row was forked from (built-in id like "reporter" or
-    # another custom id). Null for agents authored from scratch. Breadcrumb
-    # only — no FK, because built-ins live in code, not a table.
-    forked_from: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-
-    created_by: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-
-    __table_args__ = (Index("idx_custom_agents_updated_at", "updated_at"),)
-
-
-class LLMProviderConfig(Base):
-    """LLM provider configuration (Anthropic, OpenAI, Ollama, ...).
-
-    Keys are not stored here — `api_key_ref` points to a secrets_manager key.
-    See infra/database/init/09_llm_providers.sql for the table definition.
-    """
-
-    __tablename__ = "llm_provider_configs"
-
-    provider_id: Mapped[str] = mapped_column(String(64), primary_key=True)
-    provider_type: Mapped[str] = mapped_column(String(32), nullable=False)
-    name: Mapped[str] = mapped_column(String(200), nullable=False)
-    base_url: Mapped[Optional[str]] = mapped_column(String(500), nullable=True)
-    api_key_ref: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    default_model: Mapped[str] = mapped_column(String(200), nullable=False)
-    is_active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    is_default: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    config: Mapped[dict] = mapped_column(JSONB, nullable=False, default=dict)
-    last_test_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    last_test_success: Mapped[Optional[bool]] = mapped_column(Boolean, nullable=True)
-    last_error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    __table_args__ = (
-        Index("idx_llm_provider_type", "provider_type"),
-        Index("idx_llm_provider_active", "is_active"),
-        # Partial unique index — enforces "one default per provider_type"
-        # for non-Docker deployments too (Base.metadata.create_all path).
-        # Mirrors the SQL in infra/database/init/07_llm_providers.sql.
-        Index(
-            "llm_provider_default_per_type",
-            "provider_type",
-            unique=True,
-            postgresql_where=text("is_default = TRUE"),
-        ),
-    )
-
-
-class AIModelConfig(Base):
-    """Per-component AI model assignment (GH #89).
-
-    Each row maps a logical component (chat_default, triage, investigation,
-    orchestrator_plan, orchestrator_review, summarization, reporting) to a
-    (provider, model) pair. Components without a row fall back to the
-    `chat_default` row; if that is missing, callers fall back to the
-    default Anthropic provider's default_model.
-
-    See infra/database/init/10_ai_model_configs.sql for the table definition.
-    """
-
-    __tablename__ = "ai_model_configs"
-
-    component: Mapped[str] = mapped_column(String(64), primary_key=True)
-    provider_id: Mapped[str] = mapped_column(
-        String(64),
-        ForeignKey("llm_provider_configs.provider_id", ondelete="RESTRICT"),
-        nullable=False,
-    )
-    model_id: Mapped[str] = mapped_column(String(200), nullable=False)
-    settings: Mapped[dict] = mapped_column(JSONB, nullable=False, default=dict)
-    updated_by: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    __table_args__ = (Index("idx_ai_model_configs_provider", "provider_id"),)
-
-
-class ThreatIndicator(Base):
-    """Global threat indicator from external feeds (Cloudforce One STIX/TAXII, etc.).
-
-    Distinct from `CaseIOC` (case-scoped). Polled by `daemon/threat_feed_poller.py`
-    and joined against finding IOCs during enrichment in `daemon/processor.py`.
-    See `infra/database/init/14_threat_indicators.sql`.
-    """
-
-    __tablename__ = "threat_indicators"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    indicator_type: Mapped[str] = mapped_column(String(32), nullable=False)
-    indicator_value: Mapped[str] = mapped_column(String(2048), nullable=False)
-    source: Mapped[str] = mapped_column(String(64), nullable=False)
-    collection_id: Mapped[Optional[str]] = mapped_column(String(128), nullable=True)
-    confidence: Mapped[Optional[float]] = mapped_column(Numeric(5, 2), nullable=True)
-    threat_level: Mapped[Optional[str]] = mapped_column(String(16), nullable=True)
-    labels: Mapped[Optional[List[str]]] = mapped_column(ARRAY(Text), nullable=True)
-    valid_from: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    valid_until: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-    raw_stix: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-    first_seen: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    last_seen: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    __table_args__ = (
-        Index("idx_threat_indicators_type_value", "indicator_type", "indicator_value"),
-        Index("idx_threat_indicators_source", "source"),
-        Index("idx_threat_indicators_last_seen", "last_seen"),
-        Index("idx_threat_indicators_valid_until", "valid_until"),
-    )
-
-
-class Conversation(Base):
-    """Cross-device, per-analyst persistent chat conversation.
-
-    The Claude.ai-style history store for the console chat dock: a
-    listable, reopenable conversation owned by an analyst. The primary key
-    IS the frontend ``session_id`` so reopening a conversation lets the
-    in-process ``SessionManager`` (and its MemPalace files) restore live
-    context and continue the same session.
-
-    This is distinct from ``llm_interaction_logs``, which remains the
-    per-API-call compliance audit log (system-of-record). Deleting a
-    conversation here never touches that audit trail.
-    """
-
-    __tablename__ = "conversations"
-
-    # = the frontend session_id (see Chat.tsx newSessionId()).
-    id: Mapped[str] = mapped_column(String(120), primary_key=True)
-    # Owner. No hard FK: DEV_MODE's mock fallback user is not persisted, so
-    # a FK would reject those rows. user-admin-default is the seeded dev id.
-    user_id: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
-    title: Mapped[Optional[str]] = mapped_column(String(200), nullable=True)
-    agent_id: Mapped[Optional[str]] = mapped_column(String(60), nullable=True)
-    model: Mapped[Optional[str]] = mapped_column(String(80), nullable=True)
-    archived: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=False, server_default="false"
-    )
-    # Denormalized for the list view (avoids COUNT per row).
-    message_count: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0, server_default="0"
-    )
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        nullable=False,
-        default=utcnow,
-        onupdate=utcnow,
-        server_default="now()",
-    )
-    # Sort key for the history list; null until the first message lands.
-    last_message_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
-
-    messages: Mapped[List["ChatMessage"]] = relationship(
-        "ChatMessage",
-        back_populates="conversation",
-        order_by="ChatMessage.seq",
-        cascade="all, delete-orphan",
-        passive_deletes=True,
-    )
-
-    __table_args__ = (
-        Index("idx_conversations_user_last_msg", "user_id", "last_message_at"),
-        Index("idx_conversations_user_archived", "user_id", "archived"),
-    )
-
-
-class ChatMessage(Base):
-    """A single message within a :class:`Conversation`.
-
-    Full-fidelity copy of what the analyst saw live: visible text, extended
-    thinking, and the tool-call chain captured from the stream. ``complete``
-    is False for assistant turns that were aborted or errored mid-stream, so
-    a reopened chat renders exactly what was on screen.
-    """
-
-    __tablename__ = "chat_messages"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    conversation_id: Mapped[str] = mapped_column(
-        String(120),
-        ForeignKey("conversations.id", ondelete="CASCADE"),
-        nullable=False,
-    )
-    # Order within the conversation (0-based, dense). Unique per conversation.
-    seq: Mapped[int] = mapped_column(Integer, nullable=False)
-    role: Mapped[str] = mapped_column(String(20), nullable=False)
-    content: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    thinking: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    tool_calls: Mapped[list] = mapped_column(JSONB, nullable=False, default=list)
-    complete: Mapped[bool] = mapped_column(
-        Boolean, nullable=False, default=True, server_default="true"
-    )
-    model: Mapped[Optional[str]] = mapped_column(String(80), nullable=True)
-    input_tokens: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0, server_default="0"
-    )
-    output_tokens: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0, server_default="0"
-    )
-    cost_usd: Mapped[float] = mapped_column(
-        Numeric(10, 6), nullable=False, default=0, server_default="0"
-    )
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=utcnow, server_default="now()"
-    )
-
-    conversation: Mapped["Conversation"] = relationship(
-        "Conversation", back_populates="messages"
-    )
-
-    __table_args__ = (
-        UniqueConstraint("conversation_id", "seq", name="uq_chat_messages_conv_seq"),
-        Index("idx_chat_messages_conversation", "conversation_id"),
-    )
