+++ core/storage/models/__init__.py
@@ -1,0 +1,106 @@
+"""SQLAlchemy ORM models for Vigil SOC.
+
+Domain modules live beside the matching schemas in
+``core.storage.schemas``. Public names are re-exported here so
+``from core.storage.models import ...`` stays valid.
+"""
+
+from core.storage.models.ai import (
+    AIDecisionLog,
+    AIModelConfig,
+    ChatMessage,
+    Conversation,
+    LLMInteractionLog,
+    LLMProviderConfig,
+)
+from core.storage.models.auth import Role, User
+from core.storage.models.base import Base, JSONBList, case_findings
+from core.storage.models.case import Case
+from core.storage.models.case_entities import (
+    CaseAttachment,
+    CaseAuditLog,
+    CaseClosureInfo,
+    CaseComment,
+    CaseEscalation,
+    CaseEvidence,
+    CaseIOC,
+    CaseMetrics,
+    CaseNotification,
+    CaseRelationship,
+    CaseSLA,
+    CaseTask,
+    CaseTemplate,
+    CaseWatcher,
+    SLAPolicy,
+)
+from core.storage.models.config import (
+    AttackLayer,
+    ConfigAuditLog,
+    FederationSource,
+    IntegrationConfig,
+    SharedIOC,
+    SketchMapping,
+    SystemConfig,
+    ThreatIndicator,
+    UserPreference,
+)
+from core.storage.models.finding import Finding, FindingMitrePrediction
+from core.storage.models.workflow import (
+    ApprovalAction,
+    CustomAgent,
+    CustomWorkflow,
+    Investigation,
+    InvestigationLog,
+    Skill,
+    WorkflowRun,
+    WorkflowRunPhase,
+)
+
+__all__ = [
+    "AIDecisionLog",
+    "AIModelConfig",
+    "ApprovalAction",
+    "AttackLayer",
+    "Base",
+    "Case",
+    "CaseAttachment",
+    "CaseAuditLog",
+    "CaseClosureInfo",
+    "CaseComment",
+    "CaseEscalation",
+    "CaseEvidence",
+    "CaseIOC",
+    "CaseMetrics",
+    "CaseNotification",
+    "CaseRelationship",
+    "CaseSLA",
+    "CaseTask",
+    "CaseTemplate",
+    "CaseWatcher",
+    "ChatMessage",
+    "ConfigAuditLog",
+    "Conversation",
+    "CustomAgent",
+    "CustomWorkflow",
+    "FederationSource",
+    "Finding",
+    "FindingMitrePrediction",
+    "IntegrationConfig",
+    "Investigation",
+    "InvestigationLog",
+    "JSONBList",
+    "LLMInteractionLog",
+    "LLMProviderConfig",
+    "Role",
+    "SLAPolicy",
+    "SharedIOC",
+    "SketchMapping",
+    "Skill",
+    "SystemConfig",
+    "ThreatIndicator",
+    "User",
+    "UserPreference",
+    "WorkflowRun",
+    "WorkflowRunPhase",
+    "case_findings",
+]
